requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
llama-index
openai
pypdf